Added _redirects to public folder
- Also refactor code structure.
