+++ public/_redirects
@@ -1,0 +1,1 @@
+/*    /  200
